Changed initial_tree_values to long int.

+++ driver.c
@@ -7,10 +7,11 @@ int main(int argc, char **argv)
     size_t size = sizeof(int);
     size_t limit = 7;
     size_t i;
-    size_t initial_tree_values[] = { 50, 30, 20, 40, 70, 60, 80 };
+    long int initial_tree_values[] = { 50, 30, 20, 40, 70, 60, 80 };
     bst_node *root = NULL;
+
     for (i = 0; i < limit; i++) {
-        printf("Inserting into tree value: %lu\n", initial_tree_values[i]);
+        printf("Inserting into tree value: %ld\n", initial_tree_values[i]);
         root = bst_int_insert(root, size, CastInt initial_tree_values[i]);
     }
